Find optimal orders (lag structure) for ARDL bounds test

Computes optimal orders (lag structure) for the short-run relationships and autoregressive part of the ARDL model prior to ARDL bounds test with the approach of Pesaran et al. (2001).

ardlBoundOrders(data = NULL , formula = NULL, ic = c("AIC", "BIC", "MASE", 
               "GMRAE"), max.p = 15,  max.q = 15, FullSearch = FALSE  )

Arguments

data

A data.frame including all dependent and independent series. Column names this data.frame must match the variable names in formula.

formula

A formula object showing the dependent and independent series.

ic

Information criterion to be used in the serach for optimal orders.

max.p

Maximum order for the short-run coefficients.

max.q

Maximum auto-regressive order.

FullSearch

If TURE, a search over all possbilble models is implemented. See the details.

Details

If FullSearch = FALSE, this function first assumes that all p-orders are equal for the short-run relationships and finds the optimal p-order and autoregressive orders. Then, it finds the best subset of p-orders allowing them to change for each series in the short-run relationship part of the ARDL model under alternative hypothesis of ARDL bounds test.

Value

p

An integer or data.frame object including p-orders for the short-run relationship part.

q

The autoregressive order.

Stat.table

The selected statistic of all the considered models where all short-run relationship orders (p-orders) are equal.

Stat.p

The selected statistic of all possible combinations of short-run relationship orders (p-orders). The reported lag structure is the one that gives the minimum value to the chosen statistic among these combinations.

Author

Haydar Demirhan

Maintainer: Haydar Demirhan <haydar.demirhan@rmit.edu.au>