Implement finite autoregressive distributed lag model

Applies autoregressive distributed lag models of order (p , q) with one predictor.

ardlDlm(formula = NULL , data = NULL , x = NULL , y = NULL , p = 1 , q = 1 , 
         remove = NULL , type = NULL )

Arguments

formula

A formula object for the model to be fitted. In the case of multiple predictor series, the model should be entered via a formula object.

data

A data.frame including all dependent and independent series. In the case of multiple predictor series, the data should be entered via the data argument.

x

A vector including the observations of predictor time series. This is not restricted to ts objects.

y

A vector including the observations of dependent time series. This is not restricted to ts objects.

p

An integer representing finite lag length.

q

An integer representing the order of autoregressive process.

remove

A list object having two elements showing the lags of independent series with p and the autoregressive lags with q to be removed from the full model for each independent series. Please see the details for the construction of this argument.

type

An integer taking 1 if only x and y vectors are entered, 2 if a formula and data matrix is entered. It can be left NULL since the correct value is checked and fixed by the code.

Details

The autoregressive DLM is a flexible and parsimonious infinite distributed lag model. The model ARDL\((p, q)\) is written as

$$ Y_{t} = \mu+ \beta_{0}X_{t}+\beta_{1}X_{t-1}+\cdots +\beta_{p}X_{t-p}+\gamma_{1}Y_{t-1}+\cdots+\gamma_{q}Y_{t-q}+e_{t}. $$

When there is only one predictor series, both of model and formula objects can be used. But when they are supplied, both x and y arguments should be NULL.

The variable names in formula must match with the names of variables in data argument and it must be in the form of a generic formula for R functions.

The argument data contains dependent series and independent series.

The argument remove = list(p = list() , q = c()) is used to specify which lags of each independent series and the autoregressive lags of dependent series will be removed from the full model. Each element of the list p shows particular lags that will be removed from each independent series. To remove the main series from the model or to fit a model ARDL(0,q), include 0 within the elements of p. The element q is just a vector showing the autoregressive lags of dependent series to be removed.

To remove the intercept from the model, if a formula is entered, just include "-1" in the model formula. Otherwise, include "-1" in the element q of the list remove. See the examples below for implementation details.

The standard function summary() prints model summary for the model of interest.

Value

model

An object of class lm. See the details of lm function.

order

A vector composed of \(p\) and \(q\) orders.

removed.p

A list or vector showing the lags of independent series to be removed from the full model.

removed.q

A vector showing the autoregressive lags to be removed from the full model.

formula

Model formula of the fitted model. This is returned if multiple independent series are entered.

data

A data.frame including all dependent and independent series. This is returned if multiple independent series are entered.

References

B.H. Baltagi. Econometrics, Fifth Ed. Springer, 2011.

R.C. Hill, W.E. Griffiths, G.G. Judge. Undergraduate Econometrics. Wiley, 2000.

Author

Haydar Demirhan

Maintainer: Haydar Demirhan <haydar.demirhan@rmit.edu.au>

Examples

# Only one independent series

data(seaLevelTempSOI)
model.ardl = ardlDlm(x = seaLevelTempSOI$LandOcean,  
                     y = seaLevelTempSOI$GMSL, p = 1 , q = 1 )
summary(model.ardl)
#> 
#> Time series regression with "ts" data:
#> Start = 2, End = 1595
#> 
#> Call:
#> dynlm(formula = as.formula(model.text), data = data, start = 1)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -16.6769  -1.7124   0.0111   1.6847  11.8901 
#> 
#> Coefficients:
#>              Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  0.008343   0.069762   0.120    0.905    
#> X.t          1.094192   0.753292   1.453    0.147    
#> X.1         -0.750696   0.753525  -0.996    0.319    
#> Y.1          0.730866   0.017086  42.776   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 2.785 on 1590 degrees of freedom
#> Multiple R-squared:  0.5366,	Adjusted R-squared:  0.5357 
#> F-statistic: 613.7 on 3 and 1590 DF,  p-value: < 2.2e-16
#> 
# residuals(model.ardl)
# coef(model.ardl)
# fitted(model.ardl)

# Remove some lags
rem.p = c(0,1) # 0 removes the main effect of X.t
rem.q = c(1,3) 
remove = list(p = rem.p , q = rem.q)
model.ardl = ardlDlm(x = seaLevelTempSOI$LandOcean,  
                     y = seaLevelTempSOI$GMSL, p = 2 , q = 3 , remove = remove)
summary(model.ardl)
#> 
#> Time series regression with "ts" data:
#> Start = 3, End = 1595
#> 
#> Call:
#> dynlm(formula = as.formula(model.text), data = data, start = 1)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -21.516  -2.546   0.069   2.417  18.160 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  0.01554    0.09751   0.159    0.873    
#> X.2         -1.52782    1.03038  -1.483    0.138    
#> Y.2          0.30501    0.02386  12.784   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 3.892 on 1590 degrees of freedom
#> Multiple R-squared:  0.09353,	Adjusted R-squared:  0.09239 
#> F-statistic: 82.03 on 2 and 1590 DF,  p-value: < 2.2e-16
#> 

# To remove intercept as well
rem.q = c(1,3,-1) 
remove = list(p = rem.p , q = rem.q)
model.ardl = ardlDlm(x = seaLevelTempSOI$LandOcean,  
                     y = seaLevelTempSOI$GMSL, p = 2 , q = 3 , remove = remove)
summary(model.ardl)
#> 
#> Time series regression with "ts" data:
#> Start = 3, End = 1595
#> 
#> Call:
#> dynlm(formula = as.formula(model.text), data = data, start = 1)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -21.5001  -2.5305   0.0846   2.4329  18.1754 
#> 
#> Coefficients:
#>     Estimate Std. Error t value Pr(>|t|)    
#> X.t -1.52793    1.03006  -1.483    0.138    
#> Y.2  0.30502    0.02385  12.788   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 3.89 on 1591 degrees of freedom
#> Multiple R-squared:  0.09353,	Adjusted R-squared:  0.09239 
#> F-statistic: 82.08 on 2 and 1591 DF,  p-value: < 2.2e-16
#> 

# Multiple independent series
data(M1Germany)
data = M1Germany[1:144,]
model.ardlDlm  = ardlDlm(formula = logprice ~ interest + logm1, 
                         data = data.frame(data) , p = 2 , q = 1 )
summary(model.ardlDlm)
#> 
#> Time series regression with "ts" data:
#> Start = 3, End = 144
#> 
#> Call:
#> dynlm(formula = as.formula(model.text), data = data)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -0.041020 -0.011156  0.000871  0.010844  0.048895 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) -0.65648    0.24396  -2.691 0.008033 ** 
#> interest.t  -0.57167    0.33220  -1.721 0.087586 .  
#> interest.1   1.55063    0.48705   3.184 0.001808 ** 
#> interest.2  -0.41378    0.37125  -1.115 0.267032    
#> logm1.t      0.24378    0.05038   4.839 3.54e-06 ***
#> logm1.1      0.06597    0.03594   1.836 0.068641 .  
#> logm1.2     -0.19393    0.04891  -3.965 0.000119 ***
#> logprice.1   0.91192    0.03067  29.736  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.01829 on 134 degrees of freedom
#> Multiple R-squared:  0.9981,	Adjusted R-squared:  0.998 
#> F-statistic: 1.017e+04 on 7 and 134 DF,  p-value: < 2.2e-16
#> 

# To remove intercept as well
model.ardlDlm  = ardlDlm(formula = logprice ~ -1 + interest + logm1, 
                         data = data.frame(data) , p = 2 , q = 1 )
summary(model.ardlDlm)
#> 
#> Time series regression with "ts" data:
#> Start = 3, End = 144
#> 
#> Call:
#> dynlm(formula = as.formula(model.text), data = data)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -0.039495 -0.011437  0.001063  0.011899  0.045215 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> interest.t -0.427368   0.335339  -1.274  0.20470    
#> interest.1  1.574215   0.498097   3.160  0.00194 ** 
#> interest.2 -0.865766   0.338647  -2.557  0.01168 *  
#> logm1.t     0.202837   0.049124   4.129 6.34e-05 ***
#> logm1.1     0.012002   0.030504   0.393  0.69460    
#> logm1.2    -0.213138   0.049489  -4.307 3.16e-05 ***
#> logprice.1  0.993012   0.005823 170.547  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.01871 on 135 degrees of freedom
#> Multiple R-squared:      1,	Adjusted R-squared:      1 
#> F-statistic: 9.981e+05 on 7 and 135 DF,  p-value: < 2.2e-16
#> 

rem.p = list(interest = c(0,2) , logm1 = c(0))
# Remove the main series of interest and logm1 and the secont lag of 
# interest from the model
rem.q = c(1)
remove = list(p = rem.p , q = rem.q)
remove
#> $p
#> $p$interest
#> [1] 0 2
#> 
#> $p$logm1
#> [1] 0
#> 
#> 
#> $q
#> [1] 1
#> 
model.ardlDlm  = ardlDlm(formula = logprice ~ interest + logm1, 
                         data = data.frame(data) , p = 2 , q = 2 , 
                         remove = remove)
summary(model.ardlDlm) 
#> 
#> Time series regression with "ts" data:
#> Start = 3, End = 144
#> 
#> Call:
#> dynlm(formula = as.formula(model.text), data = data)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -0.038864 -0.013432 -0.002327  0.012594  0.049052 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) -0.43551    0.21914  -1.987 0.048879 *  
#> interest.1   0.56226    0.14849   3.787 0.000228 ***
#> logm1.1      0.11510    0.03793   3.034 0.002885 ** 
#> logm1.2     -0.03486    0.03637  -0.959 0.339482    
#> logprice.2   0.93454    0.02724  34.305  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.02056 on 137 degrees of freedom
#> Multiple R-squared:  0.9976,	Adjusted R-squared:  0.9975 
#> F-statistic: 1.407e+04 on 4 and 137 DF,  p-value: < 2.2e-16
#>