Implementation of Time Series Regression Models with Distributed Lag Models

Provides time series regression models with one predictor using finite distributed lag models, polynomial (Almon) distributed lag models, geometric distributed lag models with Koyck transformation, and autoregressive distributed lag models. It also consists of functions for computation of h-step ahead forecasts from these models. See Demirhan (2020)(<doi:10.1371/journal.pone.0228812>) and Baltagi (2011)(<doi:10.1007/978-3-642-20059-5>) for more information.

Details

Package:	dLagM

Type:	Package

Version:	1.1.13

Date:	2023-10-02

License:	GPL-3

To implement time series regression with finite distributed lag models, use dlm function.

To implement time series regression with polynomial distributed lag models, use polyDlm function.

To implement time series regression with geometric distributed lag models with Koyck transformation, use koyckDlm function.

To implement time series regression with autoregressive distributed lag models, use ardlDlm function.

To implement ARDL Bounds test, use ardlBound function.

To produce forecasts for any of the models, use forecast function.

To summarise the results of a model fitting, use summary function.

Author

Haydar Demirhan (https://orcid.org/0000-0002-8565-4710)

Maintainer: Haydar Demirhan <haydar.demirhan@rmit.edu.au>

Acknowledgements: The author acknowledges the testing effort of of Dr. Rogerio Porto (https://orcid.org/0000-0002-6663-9531) on forecasting function.

References

B.H. Baltagi. Econometrics, Fifth Ed. Springer, 2011.

H. Demirhan. dLagM: An R package for distributed lag models and ARDL bounds testing. PLoS ONE, 15(2): e0228812, 2020. DOI: 10.1371/journal.pone.0228812.

R.C. Hill, W.E. Griffiths, G.G. Judge. Undergraduate Econometrics. Wiley, 2000.

J. Soren, A.Q. Philips. "pss: Perform bounds test for cointegration and perform dynamic simulations."

P.K. Narayan. The Saving and Investment Nexus for China: Evidence from cointegration tests. Applied Economics 37(17):1979-1990, 2005.

M.H. Pesaran, S. Yongcheol, R.J. Smith. Bounds testing approaches to the analysis of level relationships. Journal of Applied Econometrics 16(3):289-326, 2001.

See also

Examples

# --- For examples, please refer to specific functions ---